# openapi.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "BE JS Challenge",
      "version": "1.0.11"
    },
...
    "servers": [
      {
        "url": "https://localhost:5001"
      }
    ],
...
    "components": {
      "schemas": {
...
        "accountModel": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "bank_name": {
              "type": "string"
            },
            "bank_account_number": {
              "type": "string"
            },
            "balance": {
              "type": "integer"
            },
            "user_id": {
              "type": "integer"
            }
          }
        },
...
      }
    },
    "paths": {
...
      "/api/v1/accounts": {
        "post": {
          "tags": [
            "account"
          ],
          "summary": "Create an account",
          "description": "Crate a new account from registered account",
          "operationId": "createAccount",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_id": {
                      "type": "integer"
                    },
                    "bank_name": {
                      "type": "string"
                    },
                    "bank_account_number": {
                      "type": "string"
                    },
                    "balance": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "bank_name": {
                            "type": "string"
                          },
                          "bank_account_number": {
                            "type": "string"
                          },
                          "balance": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "405": {
              "description": "Error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "get": {
          "tags": [
            "account"
          ],
          "summary": "Get all accounts",
          "description": "Return a list of all registered accounts",
          "operationId": "getAccount",
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/accountModel"
                        }
                      }
                    }
                  }
                }
              }
            },
            "405": {
              "description": "Error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/accounts/{accountId}": {
        "get": {
          "tags": [
            "account"
          ],
          "summary": "Find account by ID",
          "description": "Returns a registered account",
          "operationId": "getAccountById",
          "parameters": [
            {
              "name": "accountId",
              "in": "path",
              "description": "ID of account to return",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "bank_name": {
                            "type": "string"
                          },
                          "bank_account_number": {
                            "type": "string"
                          },
                          "balance": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "405": {
              "description": "Error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "type": "string"
                      },
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "account"
          ],
          "summary": "Updates an account",
          "description": "Updates a registered user's account",
          "operationId": "updateAccount",
          "parameters": [
            {
              "name": "accountId",
              "in": "path",
              "description": "ID of account that needs to be updated",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "bank_name": {
                      "type": "string"
                    },
                    "bank_account_number": {
                      "type": "string"
                    },
                    "balance": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "bank_name": {
                            "type": "string"
                          },
                          "bank_account_number": {
                            "type": "string"
                          },
                          "balance": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "405": {
              "description": "Client Error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "type": "string"
                      },
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "account"
          ],
          "summary": "Delete an account",
          "description": "delete an account",
          "operationId": "deleteAccount",
          "parameters": [
            {
              "name": "accountId",
              "in": "path",
              "description": "User id to delete",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      },
                      "data": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "bank_name": {
                            "type": "string"
                          },
                          "bank_account_number": {
                            "type": "string"
                          },
                          "balance": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "405": {
              "description": "Error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "type": "string"
                      },
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }